Fixed Physics issues in #6
Level now reloads on loss of Ball
